Fixes in preparation for heroku
A small number of changes had to be made in order
to get sparrow on heroku. These are:

- add gunicorn to the dependencies
- get rid of `setup.py` and rely only on `Pipfile` for dependencies
- add a Procfile so that Heroku is able to kickstart sparrow
- change logging settings to getter more granular data on `heroku logs`
- disable browsable API

The second point regarding `setup.py` deserves more explanation:
it was there to separate production dependencies with _looser_
dependency pinning (>= instead of ==, for example) while Pipfile
referred to it to pin every single dependency in the tree for
development purposes.

Turns out, Heroku gets confused by this. It sees the `setup.py`
and immediately tries to install it by running 'pipenv install -e .',
something I already made! This in turn makes their Pipfile.lock
differ from what's on the repo and, thus, aborting the deployment.

I feel like we're _slightly_ worse off by hard pinning every
dependency with Pipenv but it's alright in the end.